chore: housekeeping — dead code cleanup + file structure ordering
Phase 1 — Dead / redundant code removal:
  - Remove finalizeResponse (dead function, only called from its own test)
  - Remove 6 unused parameters: lookupReversePTR.ecsOpt,
    probeAndReorder.qtype, classifyTLD.rrtype, measureIPLatency.bgCtx,
    probeAddress.bgCtx, labelsToAdd.minimisationCount
  - Remove 3 redundant int64() conversions in handler_cache.go

Phase 2 — Standardize file-level declaration order across all packages.
  Rule added to CLAUDE.md: const → var → type → func (exported before
  unexported in each category, methods grouped by receiver).

  Files reordered:
    handler/pending.go     — PendingRequests before pendingKey/pendingCall
    handler/handler.go     — queryResult after all exported types,
                             tcpKeepaliveTimeoutForProtocol moved to end
    probe/probe.go         — var nsPending moved before all types
    hijack.go              — const before type, methods grouped by receiver
    latency/prober.go      — normalizeSteps moved to unexported helpers

+++ server/handler/handler_cache.go
@@ -212,7 +212,7 @@ func (h *Handler) processQueryError(req *dns.Msg, question Question, clientReque
 			Qname: question.Name, Qtype: question.Qtype, Qclass: question.Qclass,
 			ECS: ecsOpt, DNSSECOK: clientRequestedDNSSEC,
 			Protocol: requestProtocol, Result: "error", Rcode: dns.RcodeServerFailure,
-			ResponseTime: int64(time.Since(startTime).Milliseconds()),
+			ResponseTime: time.Since(startTime).Milliseconds(),
 		})
 		return h.buildCacheResponse(req, entry, true, question, clientRequestedDNSSEC, ecsOpt, cookieOpt, clientIP, isSecureConnection, tcpKeepaliveTimeout)
 	}
@@ -244,7 +244,7 @@ func (h *Handler) processQueryError(req *dns.Msg, question Question, clientReque
 		Qname: question.Name, Qtype: question.Qtype, Qclass: question.Qclass,
 		ECS: ecsOpt, DNSSECOK: clientRequestedDNSSEC,
 		Protocol: requestProtocol, Result: "error", Rcode: dns.RcodeServerFailure,
-		ResponseTime: int64(time.Since(startTime).Milliseconds()),
+		ResponseTime: time.Since(startTime).Milliseconds(),
 		DNSSECStatus: dnssecStatus,
 	})
 	ede := edns.NewEDEOption(edeCode, "")
@@ -297,7 +297,7 @@ func (h *Handler) processQuerySuccess(req *dns.Msg, question Question, ecsOpt *e
 	h.cache.RecordRequest(cache.RequestRecord{
 		Qname: question.Name, Qtype: question.Qtype, Qclass: question.Qclass,
 		ECS: ecsOpt, DNSSECOK: clientRequestedDNSSEC,
-		Protocol: requestProtocol, Result: "miss", ResponseTime: int64(time.Since(startTime).Milliseconds()),
+		Protocol: requestProtocol, Result: "miss", ResponseTime: time.Since(startTime).Milliseconds(),
 		Rcode: dns.RcodeSuccess, Server: server, Hijack: hijack, Fallback: fallback,
 		DNSSECStatus: dnssecStatus,
 	})
